Moved ErrorMessage component to main folder
