--- src/components/CommonComponents/index.js
@@ -1,1 +1,0 @@
-export { default as ErrorMessage } from "../CommonComponents/ErrorMessage";
